Data Retriever & Attribute Processor updates + Minor refactoring (internal file changes) + Code style fixes
Updated editorconfig and fixed code style.
Implemented characteristics computations, method to check time window, extract commits & pull comments within time window, separate method for validity, geocoder rate limit exception. Minor refactoring (without many changes to file structure) includes renaming metric names, splitting retrieveData methods, coordinates in separate method. Updated logging

+++ src/CommunityData/Characteristics.cs
@@ -6,7 +6,7 @@
     public class Characteristics
     {
         public bool Structure { get; set; }
-        public float Dispersion { get; set; }
+        public double Dispersion { get; set; }
         public float Formality { get; set; }
         public float Engagement { get; set; }
         public float Longevity { get; set; }
